pyproject.toml
[project]
name = "projeto-conta"
version = "0.1.0"
description = ""
authors = [
    {name = "marconys",email = "marconys.moura@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.135.2,<0.136.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pytz (>=2026.1.post1,<2027.0)",
    "black (>=26.3.1,<27.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
